gc 0.01

+++ app/Http/Middleware/LoginMoment.php
@@ -21,10 +21,11 @@ class LoginMoment
         if (Auth::user() === null) {
             return $next($request);
         }
-        if (Auth::user()->position_id === 1) {
+
+        if ((int) Auth::user()->position_id === 1) {
             return $next($request);
         } else {
-            if ($moment->hour >= 8 && $moment->hour <= 20 && $moment->dayOfWeek !== 0)  {
+            if ((int) $moment->hour >= 8 && (int) $moment->hour <= 20 &&  (int) $moment->dayOfWeek !== 0)  {
                 return $next($request);
             } else {
                 return redirect('at_time');
